+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Fixed
 - Updated dependencies: `std`, `liquid`.
 - Improved `isPlainObject()` util function.
+- Search improvements:
+  - Numeric arguments passed to `search.pages()` are converted to numbers.
+    For example `search.pages("foo=34")`, the `34` value is converted to number (previously it was a string, returning empty values)
+  - Values with quotes are treated as strings. For example: `search.pages("foo='34'")`, the `34` value is a string.
 
 ## [1.10.1] - 2022-07-15
 ### Added